fix(chatbot): bigint casts in budget reconcile; plan doc Phase 2 status
Postgres cannot infer types for ($3 - $2) inside GREATEST — the reconcile
UPDATE failed with 'operator is not unique' and left the pessimistic
estimate in place. Found in mock-LLM e2e verification; row now reconciles
to exact billed tokens.

+++ backend/src/chatbot/limits.ts
@@ -140,7 +140,7 @@ export async function reconcileDailyBudget(
     await db.query(
       `
         UPDATE chatbot.daily_budget
-        SET tokens_reserved = GREATEST(0, tokens_reserved + ($3 - $2))
+        SET tokens_reserved = GREATEST(0, tokens_reserved + ($3::bigint - $2::bigint))
         WHERE day = $1::date
       `,
       [reservation.day, reservation.estimatedTokens, actualTokens]
